+++ docs/pages/world/modules/_meta.js
@@ -1,5 +1,6 @@
 export default {
   "keyswithvalue": "Keys with Value",
   "keysintable": "Keys in Table",
+  "erc20": "ERC-20 tokens",
   "erc721": "ERC-721 (NFT)"
 };