+++ packages/explorer/src/app/(explorer)/Providers.tsx
@@ -1,17 +1,19 @@
 "use client";
 
-import { WagmiProvider } from "wagmi";
+import { WagmiProvider, createConfig, http } from "wagmi";
 import { injected, metaMask, safe } from "wagmi/connectors";
 import { ReactNode } from "react";
+import { RainbowKitProvider, darkTheme } from "@rainbow-me/rainbowkit";
+import "@rainbow-me/rainbowkit/styles.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { createConfig, http } from "@wagmi/core";
-import { localhost } from "@wagmi/core/chains";
-import { AppStoreProvider } from "../../store";
+import { getChain } from "../../common";
+import { defaultAnvilConnectors } from "../../connectors/anvil";
 
 const queryClient = new QueryClient();
 
+const chain = getChain();
 export const wagmiConfig = createConfig({
-  chains: [localhost],
+  chains: [chain],
   connectors: [
     injected(),
     metaMask({
@@ -20,17 +22,19 @@ export const wagmiConfig = createConfig({
       },
     }),
     safe(),
+    ...defaultAnvilConnectors,
   ],
   transports: {
-    [localhost.id]: http(),
+    [chain.id]: http(),
   },
+  ssr: true,
 });
 
 export function Providers({ children }: { children: ReactNode }) {
   return (
     <WagmiProvider config={wagmiConfig}>
       <QueryClientProvider client={queryClient}>
-        <AppStoreProvider>{children}</AppStoreProvider>
+        <RainbowKitProvider theme={darkTheme()}>{children}</RainbowKitProvider>
       </QueryClientProvider>
     </WagmiProvider>
   );