+++ packages/explorer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @latticexyz/explorer
 
+## 2.2.3
+
+### Patch Changes
+
+- b9c61a9: Fixed an issue with `--worldAddress` CLI flag being incorrectly interpreted as a number rather a hex string. Additionally, added `--hostname` option for specifying the hostname on which to start the application.
+- Updated dependencies [8546452]
+  - @latticexyz/world@2.2.3
+  - @latticexyz/store-sync@2.2.3
+  - @latticexyz/common@2.2.3
+  - @latticexyz/protocol-parser@2.2.3
+  - @latticexyz/schema-type@2.2.3
+  - @latticexyz/store@2.2.3
+
 ## 2.2.2
 
 ### Patch Changes