+++ packages/store-indexer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @latticexyz/store-indexer
 
+## 2.2.3
+
+### Patch Changes
+
+- @latticexyz/store-sync@2.2.3
+- @latticexyz/block-logs-stream@2.2.3
+- @latticexyz/common@2.2.3
+- @latticexyz/protocol-parser@2.2.3
+- @latticexyz/store@2.2.3
+
 ## 2.2.2
 
 ### Patch Changes